moving components around
